+++ src/main/resources/application.properties
@@ -1,4 +1,3 @@
 spring.mvc.view.prefix=/WEB-INF/jsp/
 spring.mvc.view.suffix=.jsp
 
-welcome.message=Hello Mkyong
