+++ src/main/webapp/WEB-INF/jsp/search.jsp
@@ -1,16 +1,13 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
-    <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-    <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Insert title here</title>
+<title>Stock News Search</title>
 </head>
 <body>
-
-
 
 <form:form method="POST" commandName="stock" action="search-news">
 <table>
@@ -19,7 +16,7 @@
         <ul>
             <form:select path="selectedStock">
                 <form:option value="-" label="--Select stock"/>
-                <form:options items="${stockMap}" />
+                <form:options items="${stockOptions}" />
             </form:select>
         </ul>
     </td>
